// src/app/home/models/news.model.ts
export class NewsModel {
  uri: string = '';
  url: string = '';
  id: number = 0;
  asset_id: number = 0;
  source: string = '';
  published_date: string = '';
  updated: string = '';
  section: string = '';
  subsection: any;
  nytdsection: string = '';
  adx_keywords: string = '';
  column: any;
  byline: string = '';
  type: string = '';
  title: string = '';
  abstract: string = '';
  des_facet: any;
  org_facet: any;
  per_facet: any;
  geo_facet: any;
  media: any;
  eta_id: number = 0;
}